Indexing Commands
Atempt number 2 at indexing commands.

These are prototype commands to test the concept. If the concept works they have a few obvious improvements.

Concept is to shove the balls into the shooter while intaking, then after intaking is done, recover the balls from the shooter so they are prepared for shooting.