Case 3 Auto Routine with PathWeaver
Worked with a few students to create an AutoRoutine that would use PathWeaver.

This path will go pick a ball, shoot 2 balls, drive to HP station with arms up, drop arms on top of ball while not moving and intake the ball, drive towards hub and wait for teleop to score.

Would be great to try to run it in the morning. I tried a few times at MVS, came home and made some fixes that I think may make it work (came home to let Brian sleep).