Added CAN_TIMEOUT to setting drivetrain encoders
The can timeout was defaulting to 0.

It says this in the comments by CTRE:
"Timeout value in ms. If nonzero, function will wait for config success and report an error if it times out. If zero, no blocking or checking is performed."

I'm worried that the no checking means no error check which means the RoboRio is not waiting to check if the encoder is actually there or not.

Added the usual 100ms timeout.